实现了消息的rollBack 以及消息进入deadQueue

+++ timeout-center-type/src/main/java/cn/bugstack/type/utils/TimeUtils.java
@@ -1,8 +1,8 @@
 package cn.bugstack.type.utils;
 
 public class TimeUtils {
-    public static double getSecondTimestamp() {
-        return (double) System.currentTimeMillis() / 1000;
+    public static long getSecondTimestamp() {
+        return System.currentTimeMillis() / 1000;
     }
 
 }
